Create SqlDelight implementation for UserPersistence

+++ server/src/main/kotlin/io/vitalir/kotlinvcshub/server/user/domain/model/UserId.kt
@@ -1,3 +1,3 @@
 package io.vitalir.kotlinvcshub.server.user.domain.model
 
-typealias UserId = Long
+typealias UserId = Int
